requirements.txt
altgraph==0.17.4
macholib==1.16.3
packaging==24.2
pillow==11.1.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
setuptools==75.8.0
tqdm==4.67.1
